Merged PR 33028: Enable automatic Win2D .winmd harvsting for MSIX manifest
## Why is this change being made?

For modern .NET platforms, the .winmd file is not directly referenced. This means that the MSIX tooling can no longer automatically harvest all its activatable classes to include in the manifest. This causes a performance hit on UWP, and it makes things not work at all in some scenarios. This PR fixes both issues.

## What changed?

Follow up to [this MSIX tooling PR](https://dev.azure.com/microsoft/ProjectReunion/_git/WindowsAppSDKClosed/pullrequest/11633978). This PR defines the "WindowsMetadataReference" needed for the MSIX tooling to automatically find and process the Win2D .winmd file. This is only added when the .winmd is not actually referenced (since in that case, things already work, and this is not needed).

Also updated the .NET projections and CsWinRT to the latest, while at it.

## Work item

https://microsoft.visualstudio.com/OS/_workitems/edit/54568138

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request enables automatic harvesting of Win2D `.winmd` files for MSIX manifests.
- `/build/nuget/Win2D.uwp-cswinrt.targets`: Added `WindowsMetadataReference` item group for Win2D `.winmd` files.
- `/build/nuget/Win2D.uwp-cswinrt.props`: New file defining properties for Win2D `.winmd` path and implementation DLL name.
- `/winrt/projection/winrt.projection.csproj`: Updated `WindowsSdkPackageVersion` and `Microsoft.Windows.CSWinRT` package version.
- `/build/nuget/Win2D.uwp.nuspec`: Included new `.props` file in the package.